Aggiunta test controller ListaBasi

+++ SitoForzeArmate/src/main/java/com/progetto/sitoforzearmate/controller/ListaBasi.java
@@ -39,7 +39,6 @@ public class ListaBasi {
         @CookieValue(value = "loggedUser", defaultValue = "") String cookieUser
         ){
             ModelAndView page = new ModelAndView();
-            DAOFactory sessionDAOFactory= null;
             DAOFactory daoFactory = null;
 
             UtenteRegistrato loggedUser = null;
@@ -48,18 +47,16 @@ public class ListaBasi {
             String applicationMessage = null;
 
             try {
-                sessionDAOFactory = DAOFactory.getDAOFactory(Configuration.COOKIE_IMPL, response);
-                sessionDAOFactory.beginTransaction();
+                if(cookieAdmin.equals("") && cookieUser.equals(""))
+                    throw new RuntimeException("Utente non loggato");
+                if(!cookieAdmin.equals("") && !cookieUser.equals(""))
+                    throw new RuntimeException("Entrambi gli utenti non possono essere loggati");
 
-                UtenteRegistratoDAO sessionUserDAO = sessionDAOFactory.getUtenteRegistratoDAO();
-                if( ! cookieUser.equals("") )    
+                if( ! cookieUser.equals("") )
                     loggedUser = UtenteRegistratoDAOcookie.decode(cookieUser);
 
-                AmministratoreDAO sessionAdminDAO = sessionDAOFactory.getAmministratoreDAO();
                 if( ! cookieAdmin.equals("") )    
                     loggedAdmin = AmministratoreDAOcookie.decode(cookieAdmin);
-
-                sessionDAOFactory.commitTransaction();
 
                 daoFactory = DAOFactory.getDAOFactory(Configuration.DAO_IMPL,null);
                 daoFactory.beginTransaction();
@@ -78,10 +75,9 @@ public class ListaBasi {
                 page.setViewName("ListaBasi/Lista");
 
             } catch (Exception e) {
-                if (sessionDAOFactory != null) sessionDAOFactory.rollbackTransaction();
-
                 e.printStackTrace();
-                page.setViewName("Pagina_InizialeCSS");
+
+                throw new RuntimeException(e);
             }
 
             return page;
@@ -94,10 +90,9 @@ public class ListaBasi {
         @CookieValue(value = "loggedUser", defaultValue = "") String cookieUser,
         @CookieValue(value = "loggedAdmin", defaultValue = "") String cookieAdmin,
         
-        @RequestParam(value = "luogoBase") String locazione
+        @RequestParam(value = "luogoBase", defaultValue = "") String locazione
         ){
             ModelAndView page = new ModelAndView();
-            DAOFactory sessionDAOFactory= null;
             DAOFactory daoFactory = null;
 
             UtenteRegistrato loggedUser = null;
@@ -106,22 +101,21 @@ public class ListaBasi {
             String applicationMessage = null;
 
             try {
-                sessionDAOFactory = DAOFactory.getDAOFactory(Configuration.COOKIE_IMPL,response);
-                sessionDAOFactory.beginTransaction();
+                if(cookieAdmin.equals("") && cookieUser.equals(""))
+                    throw new RuntimeException("Utente non loggato");
+                if(!cookieAdmin.equals("") && !cookieUser.equals(""))
+                    throw new RuntimeException("Entrambi gli utenti non possono essere loggati");
 
-                UtenteRegistratoDAO sessionUserDAO = sessionDAOFactory.getUtenteRegistratoDAO();
                 if( ! cookieUser.equals("") )
                     loggedUser = UtenteRegistratoDAOcookie.decode(cookieUser);
 
-                AmministratoreDAO sessionAdminDAO = sessionDAOFactory.getAmministratoreDAO();
                 if( ! cookieAdmin.equals("") )
                     loggedAdmin = AmministratoreDAOcookie.decode(cookieAdmin);
-
-                sessionDAOFactory.commitTransaction();
 
                 daoFactory = DAOFactory.getDAOFactory(Configuration.DAO_IMPL,null);
                 daoFactory.beginTransaction();
 
+                if(locazione.equals("")) throw new RuntimeException("Base non selezionata");
                 BaseDAO baseDAO = daoFactory.getBaseDAO();
                 Base BaseSelezionata = baseDAO.findbyLocazione(locazione);
 
@@ -135,10 +129,8 @@ public class ListaBasi {
                 page.setViewName("ListaBasi/viewBase");
 
             } catch (Exception e) {
-                if (sessionDAOFactory != null) sessionDAOFactory.rollbackTransaction();
-                
                 e.printStackTrace();
-                page.setViewName("Pagina_InizialeCSS");
+                throw new RuntimeException(e);
             }
 
             return page;
@@ -150,10 +142,9 @@ public class ListaBasi {
         
         @CookieValue(value = "loggedAdmin", defaultValue = "") String cookieAdmin,
         
-        @RequestParam(value = "luogoBase") String locazione 
+        @RequestParam(value = "luogoBase", defaultValue = "") String locazione
         ){
             ModelAndView page = new ModelAndView();
-            DAOFactory sessionDAOFactory= null;
             DAOFactory daoFactory = null;
 
             Amministratore loggedAdmin = null;
@@ -161,42 +152,33 @@ public class ListaBasi {
             String applicationMessage = null;
 
             try {
-
-                sessionDAOFactory = DAOFactory.getDAOFactory(Configuration.COOKIE_IMPL,response);
-                sessionDAOFactory.beginTransaction();
-
-                AmministratoreDAO sessionAdminDAO = sessionDAOFactory.getAmministratoreDAO();
                 if( ! cookieAdmin.equals("") )
                     loggedAdmin = AmministratoreDAOcookie.decode(cookieAdmin);
-
-                sessionDAOFactory.commitTransaction();
+                else throw new RuntimeException("Utente non loggato");
 
                 daoFactory = DAOFactory.getDAOFactory(Configuration.DAO_IMPL,null);
                 daoFactory.beginTransaction();
 
                 BaseDAO baseDAO = daoFactory.getBaseDAO();
 
-                Base base = baseDAO.findbyLocazione(locazione);   // vado a recuperare la base selezionata tramite locazione
+                if(locazione.equals("")) throw new RuntimeException("Base non selezionata");
+
+                Base base = baseDAO.findbyLocazione(locazione);         // vado a recuperare la base selezionata tramite locazione
 
                 baseDAO.delete(base);                                   // lo passo poi al metodo delete per cancellarlo
 
                 daoFactory.commitTransaction();
 
-                /*
-                page.addObject("loggedOn",loggedUser!=null);  // loggedUser != null: attribuisce valore true o false
-                page.addObject("loggedUser", loggedUser);
-                */
-                page.addObject("loggedAdminOn", loggedAdmin != null);
+                page.addObject("loggedAdminOn", true);
                 page.addObject("loggedAdmin", loggedAdmin);
                 page.addObject("BaseSelezionata", base);
 
                 page.setViewName("ListaBasi/reload");
 
             } catch (Exception e) {
-                if (sessionDAOFactory != null) sessionDAOFactory.rollbackTransaction();
-                
+
                 e.printStackTrace();
-                page.setViewName("Pagina_InizialeCSS");
+                throw new RuntimeException(e);
             }
 
             return page;
@@ -214,15 +196,16 @@ public class ListaBasi {
         try {
             if( ! cookieAdmin.equals("") )
                 loggedAdmin = AmministratoreDAOcookie.decode(cookieAdmin);
+            else throw new RuntimeException("Utente non loggato");
 
-            page.addObject("loggedAdminOn", loggedAdmin != null);
+            page.addObject("loggedAdminOn", true);
             page.addObject("loggedAdmin", loggedAdmin);
             page.setViewName("ListaBasi/NewBaseCSS");
 
         }
         catch (Exception e) {
             e.printStackTrace();
-            page.setViewName("Pagina_InizialeCSS");
+            throw new RuntimeException(e);
         }
         return page;
     }
@@ -234,51 +217,49 @@ public class ListaBasi {
         @CookieValue(value = "loggedAdmin", defaultValue = "") String cookieAdmin,
         
         @RequestParam(value = "Foto") Part foto,
-        @RequestParam(value = "Email") String email,
-        @RequestParam(value = "Telefono") String telefono,
-        @RequestParam(value = "Locazione") String locazione,
-        @RequestParam(value = "Provincia") String provincia,
-        @RequestParam(value = "CAP") String CAP,
-        @RequestParam(value = "Via") String via,
-        @RequestParam(value = "Latitudine") String latitudine,
-        @RequestParam(value = "Longitudine") String longitudine
+        @RequestParam(value = "Email", defaultValue = "") String email,
+        @RequestParam(value = "Telefono", defaultValue = "") String telefono,
+        @RequestParam(value = "Locazione", defaultValue = "") String locazione,
+        @RequestParam(value = "Provincia", defaultValue = "") String provincia,
+        @RequestParam(value = "CAP", defaultValue = "") String CAP,
+        @RequestParam(value = "Via", defaultValue = "") String via,
+        @RequestParam(value = "Latitudine", defaultValue = "") String latitudine,
+        @RequestParam(value = "Longitudine", defaultValue = "") String longitudine
         ){
             ModelAndView page = new ModelAndView();
-            DAOFactory sessionDAOFactory= null;
             DAOFactory daoFactory = null;
             Amministratore loggedAdmin = null;
             String applicationMessage = null;
 
             try {
-                sessionDAOFactory = DAOFactory.getDAOFactory(Configuration.COOKIE_IMPL,response);
-                sessionDAOFactory.beginTransaction();
-
-                AmministratoreDAO sessionAdminDAO = sessionDAOFactory.getAmministratoreDAO();
                 if( ! cookieAdmin.equals("") )
                     loggedAdmin = AmministratoreDAOcookie.decode(cookieAdmin);
+                else throw new RuntimeException("Utente non loggato");
 
                 daoFactory = DAOFactory.getDAOFactory(Configuration.DAO_IMPL,null);
                 daoFactory.beginTransaction();
 
-                /* Recupero Parametri */
+                if(foto == null || email.equals("") || telefono.equals("") || locazione.equals("") ||
+                    provincia.equals("") || CAP.equals("") || via.equals("") || latitudine.equals("") || longitudine.equals(""))
+                    throw new RuntimeException("Parametri obbligatori per la creazinoe base non forniti");
 
                 BaseDAO baseDAO = daoFactory.getBaseDAO();
                 baseDAO.create( foto.getInputStream().readAllBytes(), latitudine, longitudine, locazione, email, telefono, provincia, CAP,
                         via, loggedAdmin.getIdAdministrator());
 
                 daoFactory.commitTransaction();
-                sessionDAOFactory.commitTransaction();
 
-                page.addObject("loggedAdminOn",loggedAdmin!=null);
+                page.addObject("loggedAdminOn",true);
                 page.addObject("loggedAdmin", loggedAdmin);
                 page.addObject("applicationMessage", applicationMessage);
                 page.setViewName("ListaBasi/reload");
 
 
-            } catch (Exception e) {if (daoFactory != null) daoFactory.rollbackTransaction();
+            } catch (Exception e) {
+                if (daoFactory != null) daoFactory.rollbackTransaction();
 
                 e.printStackTrace();
-                page.setViewName("Pagina_InizialeCSS");
+                throw new RuntimeException(e);
             }
 
             return page;
